Now removes unneeded packages in the dependencies list in debian/control files
In previous versions, autovala added all and every package dependency for each
header file, binary and library needed by the program. This resulted in long
dependecies list.

Now it checks the dependencies needed for each of those packages, and removes
the packages unneeded because they are a dependency for another package.