Now the package optimization also works in non-english systems. Also removed the control.base file
Due to a mistake, the package dependency optimizations for debian added in 1.4.0 didn't work in non-english systems. This has been fixed.

Also, now instead of taking a "control.base" file to use it as a template for the
new "control" file in debian package, it takes the current "control" file, so the
changes made to it will be kept. This means that, if you want to change the
description, you just have to edit the current "control" file in the "debian" folder
and the changes will be kept every time you run "autovala deb".