Added support for specifying a tag for xgettext to filter notes for translators
xgettext has support for converting a comment before a translatable string in a
note for translators. In order to do this it is a must to add the '--add-comments'
parameter to xgettext. Also it si possible to add a tag in order to filter the
comments based on the first word, thus allowing to separate "ordinary" comments
and "notes for translators" comments.

These changes allow to include this capability in autovala, so now running
'autovala po' will also add the notes for translators.